[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int A = sc.nextInt();
-        int B = sc.nextInt();
-        int C = (A+B)*(A-B);
+        long A = sc.nextInt();
+        long B = sc.nextInt();
+        long C = (A+B)*(A-B);
         System.out.println(C);
     }
 }
